+++ bocadillo/__init__.py
@@ -1,7 +1,10 @@
 from .api import API
-from .http import *
+from .errors import HTTPError
+from .media import Media
+from .middleware import Middleware
 from .recipes import Recipe
 from .staticfiles import static
-from .websocket import WebSocket, WebSocketDisconnect
+from .views import view
+from .websockets import WebSocket, WebSocketDisconnect
 
 __version__ = "0.9.0"
