+++ bocadillo/hooks.py
@@ -2,11 +2,11 @@ import inspect
 from functools import wraps
 from typing import Callable, Dict, Union, Awaitable, Type
 
+from .compat import call_async
 from .request import Request
 from .response import Response
 from .routing import HTTPRoute
 from .views import Handler, get_handlers, View
-from ..compat import call_async
 
 HookFunction = Callable[[Request, Response, dict], Awaitable[None]]
 HookCollection = Dict[HTTPRoute, HookFunction]
