+++ tests/test_http_route_parameters.py
@@ -1,7 +1,7 @@
 import pytest
 
 from bocadillo import API
-from bocadillo.http import RouteDeclarationError
+from bocadillo.routing import RouteDeclarationError
 
 
 def test_parameter_is_passed_as_argument(api: API):
